requirements.txt
bleach==3.3.0
certifi==2021.5.30
chardet==4.0.0
colorama==0.4.4
colour-text==1.0.4
docutils==0.17.1
idna==2.10
importlib-metadata==4.5.0
keyring==23.0.1
packaging==20.9
pkginfo==1.7.0
Pygments==2.9.0
pyparsing==2.4.7
pywin32-ctypes==0.2.0
readme-renderer==29.0
requests==2.25.1
requests-toolbelt==0.9.1
rfc3986==1.5.0
six==1.16.0
tqdm==4.61.1
twine==3.4.1
urllib3==1.26.5
webencodings==0.5.1
zipp==3.4.1
pyfzf